import: classes for Git and ProcessExtensions gist from RedBlueGames
Also, placed in RedBlueGames namespace and Assembly Definition.   

Got a lot of exploratory ideas for Versioning from https://blog.redbluegames.com/version-numbering-for-games-in-unity-and-git-1d05fca83022. Solutions varied in closed-source projects, and were often very project specific. While making this into a more generalized tool, it feels better and cleaner to import and build on the inspiring work.